refactor: move api routes
