+++ views/records/list_views.py
@@ -3,8 +3,6 @@ import io
 import logging
 from flask import render_template, request, jsonify, Response
 
-from db.database import get_connection
-from db.schema import get_title_field
 from db.records import get_all_records
 
 from .record_views import records_bp
@@ -32,58 +30,6 @@ def list_view(table):
     return render_template('list_view.html', request=request, **ctx)
 
 
-@records_bp.route('/api/<table>/list')
-@require_base_table
-def api_list(table):
-    """Return id and label info for records in the table."""
-    search = request.args.get('search')
-    limit = request.args.get('limit', type=int)
-
-    with get_connection() as conn:
-        cur = conn.cursor()
-        cur.execute(f"PRAGMA table_info({table})")
-        cols = [r[1] for r in cur.fetchall()]
-        if not cols:
-            return jsonify([])
-        title_field = get_title_field(table)
-        if title_field:
-            label_field = title_field
-        elif table in cols:
-            label_field = table
-        elif len(cols) > 1:
-            label_field = cols[1]
-        else:
-            label_field = cols[0]
-
-        sql = f"SELECT id, {label_field} FROM {table}"
-        params = []
-        if search is not None:
-            sql += f" WHERE {label_field} LIKE ?"
-            params.append(f"%{search}%")
-        if limit is not None:
-            sql += " LIMIT ?"
-            params.append(limit)
-        cur.execute(sql, params)
-        rows = cur.fetchall()
-
-    return jsonify([{'id': r[0], 'label': r[1]} for r in rows])
-
-
-@records_bp.route('/api/<table>/records')
-@require_base_table
-def api_records(table):
-    ctx = build_list_context(table)
-    rows = render_template('_record_rows.html', **ctx)
-    pager = render_template('_pagination.html', **ctx)
-    count = render_template('_record_count.html', **ctx)
-    filters = render_template('_filter_chips.html', **ctx)
-    ctx.update({
-        'rows_html': rows,
-        'pagination_html': pager,
-        'count_html': count,
-        'filters_html': filters,
-    })
-    return jsonify(ctx)
 
 
 @records_bp.route('/<table>/export')
